修复有符号数无符号数混用

+++ frameworks/core/components_ng/manager/drag_drop/drag_drop_manager.cpp
@@ -173,7 +173,7 @@ int32_t DragDropManager::GetWindowId()
         }
     }
 
-    windowId = container->GetWindowId();
+    windowId = static_cast<int32_t>(container->GetWindowId());
     isDragWindowSubWindow_ = true;
 
     return windowId;
